requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-community
langchain-openai
chromadb>=0.5.0
sentence-transformers>=2.5
fastapi
uvicorn[standard]
slack-bolt>=1.18
python-dotenv
atlassian-python-api

# Optional for offline LLM
# llama-cpp-python

# Dev dependencies
pytest
ruff
black

# Additional dependencies for data processing
pandas
beautifulsoup4
lxml
aiohttp
pydantic>=2.0
typing-extensions

# Future dependencies for v2.0 (Release Tracking)
# slack-sdk>=3.23.0        # For reading Slack channels
# PyGithub>=2.1.0          # GitHub API client
# githubkit>=0.10.0        # Alternative GitHub client with GraphQL
# sqlalchemy>=2.0          # ORM for release correlations
# alembic>=1.13            # Database migrations 